requirements.txt
# ============================================================================
# REQUIREMENTS.TXT - SISTEM KOPERASI SIMPAN PINJAM
# ============================================================================
# Verified compatible versions - tested on Python 3.10+
# Install: pip install -r requirements.txt
# ============================================================================

# ====================================================================
# CORE FRAMEWORK
# ====================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ====================================================================
# DATABASE & ORM
# ====================================================================
SQLAlchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
alembic==1.13.1

# ====================================================================
# AUTHENTICATION & SECURITY
# ====================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# ====================================================================
# UTILITIES & HELPERS
# ====================================================================
python-dateutil==2.8.2
pytz==2024.1
email-validator==2.1.0

# ====================================================================
# UVICORN DEPENDENCIES (for production)
# ====================================================================
httptools==0.6.1
uvloop==0.19.0; sys_platform != 'win32'
watchfiles==0.21.0
websockets==12.0

# ====================================================================
# ASGI SERVER FOR PRODUCTION
# ====================================================================
gunicorn==21.2.0

# ====================================================================
# TESTING & DEVELOPMENT
# ====================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# ====================================================================
# CODE QUALITY & LINTING
# ====================================================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# ====================================================================
# DEPENDENCIES BREAKDOWN (for reference)
# ====================================================================

# FastAPI dependencies
anyio==4.2.0
click==8.1.7
h11==0.14.0
idna==3.6
sniffio==1.3.0
starlette==0.35.1
typing-extensions==4.9.0

# Pydantic dependencies
annotated-types==0.6.0
pydantic-core==2.14.5

# SQLAlchemy dependencies
greenlet==3.0.3

# Cryptography dependencies
cffi==1.16.0
pycparser==2.21

# Passlib dependencies
# bcrypt already listed above

# Python-jose dependencies
ecdsa==0.18.0
pyasn1==0.5.1
rsa==4.9

# Testing dependencies
coverage==7.4.0
pluggy==1.4.0
iniconfig==2.0.0
packaging==23.2
certifi==2024.2.2

# ====================================================================
# PLATFORM-SPECIFIC NOTES
# ====================================================================
# Windows: bcrypt might need Visual C++ Build Tools
# Linux: Install python3-dev and libpq-dev if needed
# macOS: Should work out of the box with Xcode Command Line Tools
# ====================================================================